+++ c++/problem_05.cpp
@@ -4,14 +4,14 @@
 
 namespace problem_05 {
 
-int solve() {
-	int n = 20;
+long solve() {
+	long n = 20;
 	bool good = false;
 
 	while (!good) {
 		n += 20;
 		good = true;
-		for (int i = 11; i < 20; ++i) {
+		for (long i = 11; i < 20; ++i) {
 			if (n % i != 0) {
 				good = false;
 				break;
